# rag/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title": "RAG Index Lab Day 21", "version": "0.1.0", "description": "Локальная индексация русской Pro Git. Размеры в символах UTF-16, не токенах. Генерации LLM нет."},
  "paths": {
    "/api/v1/corpus": {"get":{"operationId":"getCorpus","responses":{"200":{"description":"Корпус","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CorpusInfo"}}}}}}},
    "/api/v1/documents": {"get":{"operationId":"getDocuments","responses":{"200":{"description":"Документы","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentInfo"}}}}}}}},
    "/api/v1/documents/{id}": {"get":{"operationId":"getDocument","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Документ","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Document"}}}}}}},
    "/api/v1/preview": {"post":{"operationId":"previewChunks","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChunkConfig"}}}},"responses":{"200":{"description":"Preview без embeddings","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChunkPreview"}}}}}}},
    "/api/v1/jobs": {
      "get":{"operationId":"getJobs","responses":{"200":{"description":"Задания","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IndexJob"}}}}}}},
      "post":{"operationId":"startJob","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChunkConfig"}}}},"responses":{"202":{"description":"Задание принято","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexJob"}}}}}}
    },
    "/api/v1/jobs/{id}": {"get":{"operationId":"getJob","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Прогресс","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexJob"}}}}}}},
    "/api/v1/indexes": {"get":{"operationId":"getIndexes","responses":{"200":{"description":"Готовые индексы","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IndexInfo"}}}}}}}},
    "/api/v1/indexes/{id}/chunks": {"get":{"operationId":"getChunks","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Чанки","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Chunk"}}}}}}}},
    "/api/v1/indexes/{id}/documents/{documentId}": {"get":{"operationId":"getIndexDocument","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"documentId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Документ сохранённого snapshot","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Document"}}}}}}},
    "/api/v1/compare": {"post":{"operationId":"compareIndexes","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["indexIds"],"properties":{"indexIds":{"type":"array","minItems":2,"maxItems":2,"items":{"type":"string"}}}}}}},"responses":{"200":{"description":"Сравнение","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexComparison"}}}}}}},
    "/api/v1/indexes/{id}/search": {"post":{"operationId":"searchIndex","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}}},"responses":{"200":{"description":"Диагностический поиск, без LLM","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResult"}}}}}}}
  },
  "components":{"schemas":{
    "SourceFile":{"type":"object","required":["path","sha256"],"properties":{"path":{"type":"string"},"sha256":{"type":"string"}}},
    "CorpusManifest":{"type":"object","required":["corpusId","repository","revision","archiveSha256","license","attribution","selectedChapters","files"],"properties":{"corpusId":{"type":"string"},"repository":{"type":"string"},"revision":{"type":"string"},"archiveSha256":{"type":"string"},"license":{"type":"string"},"attribution":{"type":"string"},"selectedChapters":{"type":"array","items":{"type":"string"}},"files":{"type":"array","items":{"$ref":"#/components/schemas/SourceFile"}},"resources":{"type":"array","items":{"$ref":"#/components/schemas/SourceFile"}}}},
    "CorpusInfo":{"type":"object","required":["snapshotId","manifest","documentCount","sectionCount","characters","words","estimatedPages","pageFormula"],"properties":{"snapshotId":{"type":"string"},"manifest":{"$ref":"#/components/schemas/CorpusManifest"},"documentCount":{"type":"integer"},"sectionCount":{"type":"integer"},"characters":{"type":"integer"},"words":{"type":"integer"},"estimatedPages":{"type":"number"},"pageFormula":{"type":"string"}}},
    "DocumentSection":{"type":"object","required":["title","start","endExclusive"],"properties":{"title":{"type":"string"},"start":{"type":"integer"},"endExclusive":{"type":"integer"}}},
    "TextBlock":{"type":"object","required":["kind","start","endExclusive","sourceLine"],"properties":{"kind":{"type":"string"},"start":{"type":"integer"},"endExclusive":{"type":"integer"},"sourceLine":{"type":"integer"}}},
    "Document":{"type":"object","required":["id","source","sourceUrl","title","chapter","text","sha256","sections","blocks"],"properties":{"id":{"type":"string"},"source":{"type":"string"},"sourceUrl":{"type":"string"},"title":{"type":"string"},"chapter":{"type":"string"},"text":{"type":"string"},"sha256":{"type":"string"},"sections":{"type":"array","items":{"$ref":"#/components/schemas/DocumentSection"}},"blocks":{"type":"array","items":{"$ref":"#/components/schemas/TextBlock"}}}},
    "DocumentInfo":{"type":"object","required":["id","title","chapter","source","characters","words","sections"],"properties":{"id":{"type":"string"},"title":{"type":"string"},"chapter":{"type":"string"},"source":{"type":"string"},"characters":{"type":"integer"},"words":{"type":"integer"},"sections":{"type":"integer"}}},
    "ChunkConfig":{"type":"object","required":["strategy","maxCharacters","overlapCharacters"],"properties":{"strategy":{"type":"string","enum":["FIXED","STRUCTURAL"]},"maxCharacters":{"type":"integer","minimum":300,"maximum":10000},"overlapCharacters":{"type":"integer","minimum":0,"maximum":3000}}},
    "Chunk":{"type":"object","required":["chunkId","documentId","source","title","section","sections","ordinal","start","endExclusive","sourceLineStart","sourceLineEnd","text","textSha256"],"properties":{"chunkId":{"type":"string"},"documentId":{"type":"string"},"source":{"type":"string"},"title":{"type":"string"},"section":{"type":"string"},"sections":{"type":"array","items":{"type":"string"}},"ordinal":{"type":"integer"},"start":{"type":"integer"},"endExclusive":{"type":"integer"},"sourceLineStart":{"type":"integer"},"sourceLineEnd":{"type":"integer"},"text":{"type":"string"},"textSha256":{"type":"string"}}},
    "ChunkMetrics":{"type":"object","required":["chunkCount","minCharacters","medianCharacters","p95Characters","maxCharacters","uniqueCharacters","indexedCharacters","coveragePercent","crossSectionChunks","splitCodeBlocks"],"properties":{"chunkCount":{"type":"integer"},"minCharacters":{"type":"integer"},"medianCharacters":{"type":"integer"},"p95Characters":{"type":"integer"},"maxCharacters":{"type":"integer"},"uniqueCharacters":{"type":"integer"},"indexedCharacters":{"type":"integer"},"coveragePercent":{"type":"number"},"crossSectionChunks":{"type":"integer"},"splitCodeBlocks":{"type":"integer"}}},
    "EmbeddingIdentity":{"type":"object","required":["model","digest","dimension","templateVersion"],"properties":{"model":{"type":"string"},"digest":{"type":"string"},"dimension":{"type":"integer"},"templateVersion":{"type":"string"}}},
    "IndexInfo":{"type":"object","required":["id","snapshotId","createdAt","config","embedding","metrics","buildMilliseconds","embeddingMilliseconds","embeddingInputTokens","vectorBytes"],"properties":{"id":{"type":"string"},"snapshotId":{"type":"string"},"createdAt":{"type":"string"},"config":{"$ref":"#/components/schemas/ChunkConfig"},"embedding":{"$ref":"#/components/schemas/EmbeddingIdentity"},"metrics":{"$ref":"#/components/schemas/ChunkMetrics"},"buildMilliseconds":{"type":"integer"},"embeddingMilliseconds":{"type":"integer"},"embeddingInputTokens":{"type":"integer","nullable":true},"vectorBytes":{"type":"integer"}}},
    "IndexJob":{"type":"object","required":["id","config","status","processed","total","createdAt","updatedAt","indexId","error"],"properties":{"id":{"type":"string"},"config":{"$ref":"#/components/schemas/ChunkConfig"},"status":{"type":"string","enum":["QUEUED","RUNNING","READY","FAILED"]},"processed":{"type":"integer"},"total":{"type":"integer"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"},"indexId":{"type":"string","nullable":true},"error":{"type":"string","nullable":true}}},
    "ChunkPreview":{"type":"object","required":["snapshotId","config","metrics","chunks"],"properties":{"snapshotId":{"type":"string"},"config":{"$ref":"#/components/schemas/ChunkConfig"},"metrics":{"$ref":"#/components/schemas/ChunkMetrics"},"chunks":{"type":"array","items":{"$ref":"#/components/schemas/Chunk"}}}},
    "IndexComparison":{"type":"object","required":["comparable","notes","indexes"],"properties":{"comparable":{"type":"boolean"},"notes":{"type":"array","items":{"type":"string"}},"indexes":{"type":"array","items":{"$ref":"#/components/schemas/IndexInfo"}}}},
    "SearchRequest":{"type":"object","required":["query","topK"],"properties":{"query":{"type":"string","minLength":1,"maxLength":2000},"topK":{"type":"integer","minimum":1,"maximum":20}}},
    "SearchHit":{"type":"object","required":["rank","similarity","chunk"],"properties":{"rank":{"type":"integer"},"similarity":{"type":"number"},"chunk":{"$ref":"#/components/schemas/Chunk"}}},
    "SearchResult":{"type":"object","required":["indexId","query","milliseconds","inputTokens","hits"],"properties":{"indexId":{"type":"string"},"query":{"type":"string"},"milliseconds":{"type":"integer"},"inputTokens":{"type":"integer","nullable":true},"hits":{"type":"array","items":{"$ref":"#/components/schemas/SearchHit"}}}},
    "ApiError":{"type":"object","required":["code","message"],"properties":{"code":{"type":"string"},"message":{"type":"string"}}}
  }}
}